fix wifi support in linuxrc (bsc #949938)
This is a complete rewrite of the wifi related code to work
with wicked.

linxurc supports open wlan, wpa-psk, and peap authentication.
If you need anything more specific, you can still specify the
required WIRELESS_* parameters via ifcfg option.

There is a new wlan.device option to tag an arbitray interface
as wifi-capable (for testing or to override auto-detection).