add support for compressed modules
We enabled compressed modules in 5.3. We added support into
installation-images, but forgot to add the support to linuxrc. So do it
now.

This removes the global hardcoded MODULE_SUFFIX and introduces a list of
possible module extensions instead. New functions mod_has_module_ext and
mod_find_module are introduced to cover the functionality. It also
cleans up the code significantly.

config.forceinsmod works differently in mod_insmod now. We have to check
the path as we have to find whether the module is compressed or not. But
I think it makes little sense not to do so. Looking at b1033964884a, I
cannot see why the check is skipped when force-loading?