read and parse autoyast file
This basically moves the existing code into a separate function.

autoyast2 still exists as a fallback to autoyast to keep existing workflows
working.

Note that the default action 'autoyast=default' is represented as 'no
autoyast url' in linuxrc. In this case it looks for an autoinst.xml file
at the repo location.